+++ examples/aliases/main.tf
@@ -1,7 +1,8 @@
 terraform {
   required_providers {
     hubspot = {
-      source = "registry.opentofu.org/jackemcpherson/hubspot"
+      source  = "registry.opentofu.org/jackemcpherson/hubspot"
+      version = "= 0.1.0-alpha.1"
     }
   }
 }